requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.11
click==7.1.2
colorama==0.4.4
emoji==1.7.0
et-xmlfile==1.1.0
future==0.18.2
greenlet==1.1.2
idna==3.3
openpyxl==3.0.9
pefile==2021.9.3
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
python-dotenv==0.19.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
tornado==6.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.8
